Refactor helper methods into Query
