+++ Test_Engine/Test_Engine.csproj
@@ -96,7 +96,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Convert\ToCSharpSyntaxTree.cs" />
+    <Compile Include="Convert\ToSyntaxTree.cs" />
     <Compile Include="Convert\ToSpan.cs" />
     <Compile Include="Convert\ToLineSpan.cs" />
     <Compile Include="Convert\ToPosition.cs" />
@@ -106,6 +106,8 @@
     <Compile Include="Create\LineSpan.cs" />
     <Compile Include="Create\Span.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Query\GetLeadingTrivia.cs" />
+    <Compile Include="Query\GetRoot.cs" />
     <Compile Include="Query\TestDataOfType.cs" />
     <Compile Include="Query\GetText.cs" />
   </ItemGroup>
